now users can comment on posts

+++ Post/views.py
@@ -2,45 +2,56 @@ from django.shortcuts import render,redirect
 from django.core.files.storage import FileSystemStorage
 from django.contrib import auth, messages
 from django.db.models import Count
-from .models import Category, Featured, Like, Post
+from .models import Category, Comment, Featured, Like, Post
 
 # Create your views here.
 
+# blog main page
 def index(request):
     top_authors = Post.objects.values('author').annotate(no_of_post=Count('id')).order_by('-no_of_post')[:3]
     top_posts = Like.objects.values('liked_on').annotate(likes=Count('liked_by')).order_by('-likes')[:3]
-    for post in top_posts:
-        top_post = Post.objects.all().filter(id=post['liked_on'])
-        print(top_post)
+    print(top_posts)
     posts = Post.objects.all().order_by('-created_at')[:3]
     featured = Featured.objects.all()
     
     return render(request, 'blog/index.html',{'posts':posts,'featured':featured,'top_posts':top_posts})
 
 
-
+# all blogs page
 def blogs(request):
-
     posts = Post.objects.all().order_by('-created_at')
-
     return render(request,'blog/blog.html',{'posts':posts})
 
 
-
+# blogs categories page
 def categories(request):
-
     categories = Category.objects.all()
-
-    return render(request,'blog/categories.html',{'categories': categories})
-
+    return render(request,'blog/categories.html',{'categories':categories})
 
 
+# filter posts by category
+def blogFilter(request, id):
+    posts = Post.objects.filter(category_id=id).order_by('-created_at')
+    return render(request,'blog/blog.html',{'posts':posts})
+
+
+# filter single post
 def single(request, id):
     post = Post.objects.get(id=id)
-    return render(request,'blog/single.html',{'post':post})
+    comments = Comment.objects.filter(comment_on=id).annotate(count=Count('id')).order_by()
+    print(comments)
+    if request.method=='POST':
+        if request.user.is_authenticated:
+            comment = request.POST['comment']
+            query = Comment(comment=comment,comment_by=request.user, comment_on=post)
+            back = request.POST.get('back', '/')
+            query.save()
+            return redirect(back)
+    related = Post.objects.filter(category_id=post.category_id).order_by('-created_at')[:3]
+    return render(request,'blog/single.html',{'post':post, 'related':related, 'comments':comments})
 
 
-
+# create post
 def create(request):
     if request.user.is_authenticated:
         if request.method == 'POST' and request.FILES['image']:
@@ -52,13 +63,14 @@ def create(request):
             image = fs.save(file.name, file)
             author = request.user.id
             post = Post(title=title, description=description, category_id=category, author_id = author, image=image)
+            print(post)
             post.save()
         categories = Category.objects.all()
-        return render(request, 'blog/create.html',{'categories': categories})
+        return render(request, 'blog/create.html',{'categories':categories})
     return redirect('index')
 
 
-
+# edit post
 def edit(request, id):
     if request.user.is_authenticated:
         post = Post.objects.get(id=id)
@@ -72,35 +84,42 @@ def edit(request, id):
                 post.image = fs.save(file.name, file)
                 post.author_id = request.user.id
                 post.save()
-                return redirect('profile')
+                return redirect('yourblog')
             else:
                 post.title = request.POST['title']
                 post.description = request.POST['description']
                 post.category_id = request.POST.get('category')
                 post.author_id = request.user.id
                 post.save()
-                return redirect('profile')
+                return redirect('yourblog')
+        if int(request.user.id) != int(post.author_id):
+            return redirect('index')
         categories = Category.objects.all()
-        return render(request, 'blog/edit.html',{'categories': categories,'post':post})
+        return render(request, 'blog/edit.html',{'categories':categories, 'post':post})
     return redirect('index')
 
 
-
+# delete post
 def delete(request, id):
     if request.user.is_authenticated:
         post = Post.objects.get(id=id)
+        if int(request.user.id) != int(post.author_id):
+            return redirect('index')
         post.delete()
-        return redirect('profile')
+        return redirect('yourblog')
     return redirect('index')
 
 
-
-def profile(request):
+# blog created by the logged in user
+def yourBlog(request):
     if request.user.is_authenticated:
-
         posts = Post.objects.filter(author_id=request.user.id).all().order_by('-created_at')
-
-        return render(request,'blog/profile.html',{'posts': posts})
-
+        return render(request,'blog/yourblog.html',{'posts':posts})
     return redirect('index')
     
+
+# profile of logged in user
+def profile(request):
+    if request.user.is_authenticated:
+        return render(request,'blog/profile.html')
+    return redirect('index')
